Ajout lecture fichier config - erreur
Erreur => La structure config n'est pas remplie
