+++ cmd/sensor/main.go
@@ -2,12 +2,12 @@ package main
 
 import (
 	"Moniport/cmd/mqtt"
+	"encoding/json"
+	"flag"
 	"fmt"
+	"os"
 	"strconv"
 	"time"
-	"flag"
-	"os"
-	"encoding/json"
 )
 
 type configuration struct {
@@ -18,6 +18,14 @@ type configuration struct {
 }
 
 func main() {
+	//Lecture de l'adresse du fichier de config => lancer "go run <adresse-main.go> -config <adresse-config.json>
+	configFilename := getArgConfig()
+	fmt.Println(configFilename)
+
+	//Lecture du fichier de configuration
+	config := readConfiguration(configFilename)
+	fmt.Println(config)
+
 	s1 := Sensor{
 		id:         001,
 		idAirport:  "NTE",
@@ -34,4 +42,34 @@ func main() {
 		c1.Publish("topic", s1.mqttQos, false, s1.GenerateMessage(time.Now()))
 	}
 
+}
+
+func getArgConfig() string {
+	var configFilename string
+	flag.StringVar(&configFilename, "config", "", "Usage")
+	flag.Parse()
+	return configFilename
+}
+
+func readConfiguration(filename string) configuration {
+
+	var _configuration configuration
+	//filename is the path to the json config file
+	file, err := os.Open(filename)
+	if err != nil {
+		fmt.Println("Error 1")
+		return _configuration
+	}
+	decoder := json.NewDecoder(file)
+	err = decoder.Decode(&_configuration)
+	if err != nil {
+		fmt.Println("Error 2")
+		return _configuration
+	}
+
+	//TODO error handling
+
+	fmt.Println("file read")
+	fmt.Println(_configuration.adressBroker)
+	return _configuration
 }
